+++ providers/site.rb
@@ -21,10 +21,10 @@
 
 action :enable do
   timing = [:delayed, :immediately].include?(new_resource.timing) ? new_resource.timing : :delayed
-  link_name = (new_resource.name == "default") ? "000-default" : new_resource.name
+  link_name = (new_resource.name == 'default') ? '000-default' : new_resource.name
   a = execute "nxensite #{new_resource.name}" do
     command "/usr/sbin/nxensite #{new_resource.name}"
-    notifies :reload, 'service[nginx]', timing
+    notifies :reload, node['openresty']['service']['resource'], timing
     not_if { ::File.symlink?("#{node['openresty']['dir']}/sites-enabled/#{link_name}") }
   end
 
@@ -33,10 +33,10 @@ end
 
 action :disable do
   timing = [:delayed, :immediately].include?(new_resource.timing) ? new_resource.timing : :delayed
-  link_name = (new_resource.name == "default") ? "000-default" : new_resource.name
+  link_name = (new_resource.name == 'default') ? '000-default' : new_resource.name
   a = execute "nxdissite #{new_resource.name}" do
     command "/usr/sbin/nxdissite #{new_resource.name}"
-    notifies :reload, 'service[nginx]', timing
+    notifies :reload, node['openresty']['service']['resource'], timing
     only_if { ::File.symlink?("#{node['openresty']['dir']}/sites-enabled/#{link_name}") }
   end