+++ recipes/commons_dir.rb
@@ -50,7 +50,6 @@ directory '/var/www' do
   action :create
 end
 
-
 %w(client_temp proxy_temp fastcgi_temp uwsgi_temp scgi_temp).each do |leaf|
   directory File.join(node['openresty']['cache_dir'], leaf) do
     owner node['openresty']['user']