+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - OpenResty version bump
 - Added automatic creation of /var/cache/nginx/{fastcgi,scgi,uwsgi,proxy}_cache
 - Added support for linking with jemalloc. May provide some benefits when used in serving LUA
+- Minor file naming fix
 
 ## 0.1.9: